feat(base): handle x-systemd.device-timeout and nofail fstab mount options
The base module adds initqueue hooks and systemd device dropins that disable
timeouts to wait indefinitely for host devices that need to be mounted in the
initrd (saved in `host_devs`, and not present in `root_devs` or `swap_devs`).
Usually these devices are added using `--add-fstab`, `--mount` or `--add-device`
command line options, or collecting entries from /etc/fstab with the
`x-initrd.mount` option.

This patch aims to add support for the `x-systemd.device-timeout` and `nofail`
fstab options for devices added via `x-initrd.mount`, `--add-fstab` or
`--mount`, so:
- If `x-systemd.device-timeout` is set, `JobTimeoutSec` and
`JobRunningTimeoutSec` will use this value, instead of 0.
- If `nofail` is set, it will be added an initqueue timeout script which calls
`cancel_wait_for_dev`, preventing emergency shell entry if the device could not
be mounted in the initrd.