fix(iscsi): keep iscsiuio optional
The iscsi module is always installed in no-hostonly mode when building a
generic initrd, issuing "Failed to enable unit, unit iscsiuio.socket does not
exist." if iscsiuio is not installed.

If iscsiuio is not installed, do not try to enable iscsiuio.socket, install
custom dracut configuration in the initrd or create a dependency in
iscsi-shutdown.service.