fix(nvmf): honor rd.timeout for nvme ctrl_loss_tmo
Connection issues during boot may cause the kernel to give up retrying
the NVMe connect command prematurely. Use "rd.timeout" to set the
ctrl_loss_tmo value for the NVMe subsystem. This means that the default
timeout will be infinite.

We use an udev rule to set the timeout. We want to do this only for controllers
that are brought up in the initramfs. Therefore write the udev rule
to /etc/udev/rules.d rather than /run/udev/rules.d.
Strictly speaking, we'd want to set the timeouot only for those controllers
that are necessary for booting, but it would be very difficult to identify
these reliably, so just use this simple approach for now.

Signed-off-by: Martin Wilck <mwilck@suse.com>