feat(suse-initrd): drop support for INITRD_MODULES
This is compat code to support the `mkinitrd` method used to force inclusion of
modules in the initrd up to SLE11. After removing the `mkinitrd` wrapper in
https://github.com/openSUSE/dracut/commit/83bf6728, it's time to remove this
legacy code.

With dracut, the native way to include modules in the initrd is the
`add_drivers` or `force_drivers` configuration options.

jsc#PED-1915